commit log and code clean up

+++ raft-server/src/main/java/com/josiahebhomenye/raft/server/core/LogEntry.java
@@ -9,13 +9,13 @@ import java.nio.ByteBuffer;
 
 @Value
 public class LogEntry {
-    private int term;
+    private long term;
     private Command command;
 
     public byte[] serialize(){
         ByteBuf buf = Unpooled.buffer();
         byte[] command = this.command.serialize();
-        buf.writeInt(term);
+        buf.writeLong(term);
         buf.writeBytes(command);
 
         byte[] content = new byte[buf.readableBytes()];
@@ -27,7 +27,7 @@ public class LogEntry {
     public static LogEntry deserialize(byte[] entry){
         ByteBuf buf = Unpooled.wrappedBuffer(entry);
 
-        int term = buf.readInt();
+        long term = buf.readLong();
         byte[] data = new byte[buf.readableBytes()];
         buf.readBytes(data);
         Command command = Command.restore(data);
